GDP calc corrections
All 3 methods now are 6-digit figures of 7 or 8 hundred thousand. Still not the same so not entirely correct. But better.
In order to do this, some variables had to be changed.
The main DF is now _transactions, and a slice of that is made for only the transactions. Change if necessary.
Also, check if intermediate calculations/code are still correct. Might not serve the intended purpose anymore.
After this, start new code file.